backend/pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.1.0"
description = "Backend Foundation for Todo Web Application - JWT Verification and Todo Management"
authors = ["Developer <developer@example.com>"]
package-mode = false

[tool.poetry.scripts]
start = "src.main:main"


[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlmodel = "^0.0.16"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
better-exceptions = "^0.3.3"
pydantic = {version = "^2.5.0", extras = ["email"]}
pydantic-settings = "^2.0.3"
typing-extensions = "^4.8.0"
neon = "^0.1.0"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true